Record entire resolution hierarchy in TimeSeries type.
Previously `TimeSeries` would only record the parent and child type, now we use
a type level list to keep track of the entire resolution hierarchy. For example,
we can now have a `TimeSeries '[ 'Year, 'Quarter, 'Month ]` if we have a years
worth of data broken down by quarter and then again by month.

We also switch to using promoted data constructors to represent the timeseries
period instead of unique types, you can see this in `data TsPeriod`.

In addition, we introduced the `BuildTS` (build time series) typeclass for
conveniently constructing timeseries with the appropriate type given the
context. This is used in helper functions such as `raw :: BuildTS p => Double ->
TimeSeries '[p]`.

Last, we implemented a simple `computeTotal` function to aggregate the data in a
timeseries and propagate it to the root.